+++ runtime/vm/compiler/backend/constant_propagator.cc
@@ -232,8 +232,10 @@ void ConstantPropagator::VisitGoto(GotoInstr* instr) {
 }
 
 void ConstantPropagator::VisitIndirectGoto(IndirectGotoInstr* instr) {
-  for (intptr_t i = 0; i < instr->SuccessorCount(); i++) {
-    SetReachable(instr->SuccessorAt(i));
+  if (reachable_->Contains(instr->GetBlock()->preorder_number())) {
+    for (intptr_t i = 0; i < instr->SuccessorCount(); i++) {
+      SetReachable(instr->SuccessorAt(i));
+    }
   }
 }
 
@@ -1502,7 +1504,11 @@ static bool HasPhis(BlockEntryInstr* block) {
 }
 
 static bool IsEmptyBlock(BlockEntryInstr* block) {
-  return block->next()->IsGoto() && !HasPhis(block) &&
+  // A block containing a goto to itself forms an infinite loop.
+  // We don't consider this an empty block to handle the edge-case where code
+  // reduces to an infinite loop.
+  return block->next()->IsGoto() &&
+         block->next()->AsGoto()->successor() != block && !HasPhis(block) &&
          !block->IsIndirectEntry();
 }