+++ runtime/vm/compiler/backend/locations.h
@@ -21,6 +21,8 @@ namespace dart {
 class BaseTextBuffer;
 class ConstantInstr;
 class Definition;
+class FlowGraphDeserializer;
+class FlowGraphSerializer;
 class PairLocation;
 class Value;
 
@@ -430,8 +432,8 @@ class Location : public ValueObject {
 
   Location Copy() const;
 
-  static Location read(uword value) { return Location(value); }
-  uword write() const { return value_; }
+  void Write(FlowGraphSerializer* s) const;
+  static Location Read(FlowGraphDeserializer* d);
 
  private:
   explicit Location(uword value) : value_(value) {}
@@ -572,8 +574,7 @@ class RegisterSet : public ValueObject {
     ASSERT(kNumberOfFpuRegisters <= (kWordSize * kBitsPerByte));
   }
 
-  explicit RegisterSet(uintptr_t cpu_register_mask,
-                       uintptr_t fpu_register_mask = 0)
+  explicit RegisterSet(uintptr_t cpu_register_mask, uintptr_t fpu_register_mask)
       : RegisterSet() {
     AddTaggedRegisters(cpu_register_mask, fpu_register_mask);
   }
@@ -717,6 +718,9 @@ class RegisterSet : public ValueObject {
     untagged_cpu_registers_.Clear();
   }
 
+  void Write(FlowGraphSerializer* s) const;
+  explicit RegisterSet(FlowGraphDeserializer* d);
+
  private:
   SmallSet<Register> cpu_registers_;
   SmallSet<Register> untagged_cpu_registers_;
@@ -834,6 +838,9 @@ class LocationSummary : public ZoneAllocated {
   void DiscoverWritableInputs();
   void CheckWritableInputs();
 #endif
+
+  void Write(FlowGraphSerializer* s) const;
+  explicit LocationSummary(FlowGraphDeserializer* d);
 
  private:
   BitmapBuilder& EnsureStackBitmap() {